[misc] Device API public headers and CMake rework part 1 (#7624)
Issue: #

### Brief Summary

Reworked the device API cmake structure and created a device API target
that contains all device API source. Seperated (temporarily) `device.h`
and `public_device.h` to isolate a publicly usable interface. Next step
is to remove currently `PUBLIC` header dependencies & dependencies on
Taichi core src.

The majority of the change set (around 900~1000 lines of changes) is
simply moving the publicly usable contents of `device.h` to
`public_device.h`

---------
