+++ taichi/common/CMakeLists.txt
@@ -11,6 +11,7 @@ target_sources(taichi_common
     virtual_dir.cpp
     zip.cpp
     miniz.c
+    dynamic_loader.cpp
   )
 
 target_include_directories(taichi_common
