+++ taichi/runtime/llvm/CMakeLists.txt
@@ -20,7 +20,6 @@ target_include_directories(llvm_runtime
     ${LLVM_INCLUDE_DIRS}
   )
 
-target_link_libraries(llvm_runtime PRIVATE llvm_rhi)
 target_link_libraries(llvm_runtime PRIVATE ${llvm_libs})
 
 
@@ -30,15 +29,14 @@ endif()
 
 if (TI_WITH_CUDA)
     target_link_libraries(llvm_runtime PRIVATE ${llvm_ptx_libs})
-    target_link_libraries(llvm_runtime PRIVATE cuda_rhi)
 endif()
 
 if (TI_WITH_AMDGPU)
     target_link_libraries(llvm_runtime PRIVATE ${llvm_amdgpu_libs})
-    target_link_libraries(llvm_runtime PRIVATE amdgpu_rhi)
 endif()
 
 if (TI_WITH_DX12)
     target_link_libraries(llvm_runtime PRIVATE ${llvm_directx_libs})
-    target_link_libraries(llvm_runtime PRIVATE dx12_rhi)
 endif()
+
+target_link_libraries(llvm_runtime PRIVATE ti_device_api)
