+++ taichi/runtime/program_impls/dx/CMakeLists.txt
@@ -15,5 +15,4 @@ target_include_directories(dx_program_impl
     ${LLVM_INCLUDE_DIRS}
   )
 
-target_link_libraries(dx_program_impl PRIVATE dx_rhi)
 target_link_libraries(dx_program_impl PRIVATE compilation_manager)
