+++ taichi/runtime/program_impls/metal/CMakeLists.txt
@@ -18,4 +18,3 @@ target_include_directories(metal_program_impl
 target_link_libraries(metal_program_impl PRIVATE spirv_codegen)
 target_link_libraries(metal_program_impl PRIVATE gfx_runtime)
 target_link_libraries(metal_program_impl PRIVATE compilation_manager)
-target_link_libraries(metal_program_impl PRIVATE metal_rhi)
