+++ taichi/runtime/program_impls/opengl/CMakeLists.txt
@@ -17,4 +17,3 @@ target_include_directories(opengl_program_impl
 
 target_link_libraries(opengl_program_impl PRIVATE gfx_runtime)
 target_link_libraries(opengl_program_impl PRIVATE compilation_manager)
-target_link_libraries(opengl_program_impl PRIVATE opengl_rhi)
