+++ taichi/runtime/program_impls/vulkan/CMakeLists.txt
@@ -4,6 +4,7 @@ add_library(vulkan_program_impl)
 target_sources(vulkan_program_impl
   PRIVATE
     vulkan_program.cpp
+    vulkan_program.h
   )
 
 target_include_directories(vulkan_program_impl
@@ -19,4 +20,5 @@ target_include_directories(vulkan_program_impl
 target_link_libraries(vulkan_program_impl PRIVATE spirv_codegen)
 target_link_libraries(vulkan_program_impl PRIVATE gfx_runtime)
 target_link_libraries(vulkan_program_impl PRIVATE compilation_manager)
+target_link_libraries(vulkan_program_impl PRIVATE ti_device_api)
 target_link_libraries(vulkan_program_impl PRIVATE vulkan_rhi)
